+++ couriercxx/connection/event/EventRead.cpp
@@ -14,40 +14,40 @@
 
 namespace Connection {
 
-EventRead::EventRead(const Info info, const uint8_t* data, uint32_t dataLen) : IEvent(Connection::EVENT_READ), EventConnection(info) {
-	if ((data != nullptr) && (dataLen > 0)) {
-		this->data = new uint8_t[dataLen];
-		this->dataLen = dataLen;
-		memmove(this->data, data, sizeof(uint8_t) * this->dataLen);
+EventRead::EventRead(const Info info, const uint8_t* data, uint32_t dataSize) : IEvent(Connection::EVENT_READ), EventConnection(info) {
+	if ((data != nullptr) && (dataSize > 0)) {
+		this->data = new uint8_t[dataSize];
+		this->dataSize = dataSize;
+		memmove(this->data, data, sizeof(uint8_t) * this->dataSize);
 	} else {
 		this->data = nullptr;
-		this->dataLen = 0;
+		this->dataSize = 0;
 	}
 }
 
-EventRead::EventRead(const Info info, const uint8_t* data, uint32_t dataLen, IListener* source, IListener* target) : IEvent(Connection::EVENT_READ, source, target), EventConnection(info) {
-	if ((data != nullptr) && (dataLen > 0)) {
-		this->data = new uint8_t[dataLen];
-		this->dataLen = dataLen;
-		memmove(this->data, data, sizeof(uint8_t) * this->dataLen);
+EventRead::EventRead(const Info info, const uint8_t* data, uint32_t dataSize, IListener* source, IListener* target) : IEvent(Connection::EVENT_READ, source, target), EventConnection(info) {
+	if ((data != nullptr) && (dataSize > 0)) {
+		this->data = new uint8_t[dataSize];
+		this->dataSize = dataSize;
+		memmove(this->data, data, sizeof(uint8_t) * this->dataSize);
 	} else {
 		this->data = nullptr;
-		this->dataLen = 0;
+		this->dataSize = 0;
 	}
 }
 
 EventRead::EventRead(const EventRead& other) : IEvent(other), EventConnection(other) {
-	dataLen = other.dataLen;
-	if (dataLen > 0) {
-		data = new uint8_t[dataLen];
-		memmove(data, &other.data, sizeof(uint8_t) * dataLen);
+	dataSize = other.dataSize;
+	if (dataSize > 0) {
+		data = new uint8_t[dataSize];
+		memmove(data, &other.data, sizeof(uint8_t) * dataSize);
 	} else {
 		data = nullptr;
 	}
 }
 
 EventRead::EventRead(EventRead&& other) : IEvent(other), EventConnection(other) {
-	dataLen = other.dataLen;
+	dataSize = other.dataSize;
 	data = other.data;
 	other.data = nullptr;
 }
@@ -64,10 +64,10 @@ EventRead& EventRead::operator=(const EventRead& other) {
 	EventConnection::operator=(other);
 
 	if (this != &other) {
-		dataLen = other.dataLen;
-		if (dataLen > 0) {
-			data = new uint8_t[dataLen];
-			memmove(data, &other.data, sizeof(uint8_t) * dataLen);
+		dataSize = other.dataSize;
+		if (dataSize > 0) {
+			data = new uint8_t[dataSize];
+			memmove(data, &other.data, sizeof(uint8_t) * dataSize);
 		} else {
 			data = nullptr;
 		}
@@ -81,7 +81,7 @@ EventRead& EventRead::operator=(EventRead&& other) {
 	EventConnection::operator=(other);
 
 	if (this != &other) {
-		dataLen = other.dataLen;
+		dataSize = other.dataSize;
 		data = other.data;
 		other.data = nullptr;
 	}
@@ -89,16 +89,16 @@ EventRead& EventRead::operator=(EventRead&& other) {
 	return *this;
 }
 
-uint32_t EventRead::getDataLen() const {
-	return dataLen;
+uint32_t EventRead::getDataSize() const {
+	return dataSize;
 }
 
 int EventRead::getData(uint8_t* buffer, uint32_t bufferSize) const {
-	if ((data == nullptr) || (buffer == nullptr) || (bufferSize == 0) || (dataLen == 0)) {
+	if ((data == nullptr) || (buffer == nullptr) || (bufferSize == 0) || (dataSize == 0)) {
 		return -1;
 	}
 
-	uint32_t minLen = bufferSize > dataLen ? dataLen : bufferSize;
+	uint32_t minLen = bufferSize > dataSize ? dataSize : bufferSize;
 	memmove(buffer, data, sizeof(uint8_t) * minLen);
 
 	return 0;