+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Version 0.4.0.0
+[*] dataLen -> dataSize
+
 ### Version 0.3.0.0
 [+] added initial Windows support